More complete with a menu bar and improved functionnality

+++ file.h
@@ -5,7 +5,7 @@
 
 struct ServerFile {
     QString filePath;
-    qint32 fileSize;
+    qint64 fileSize;
     qint16 numberDownloads;
 };
